[OGLEngine] experimented with updating extents each frame, based on the min and max x,y,z values out of all vertices, so extents are not tied to two specific vertices but the actual min and max x,y,z values out of all of them. It feels right in my gut with how the extents are calculated now as a result of this.
Regarding if this improves the SphereBoxTest, it may be so. Both position and scale of the box works with the sphere, meaning I can move the box around the sphere and scale it however I want and collisions seem correct. I couldn't do that before, where the collision would be offset whenever I moved the Box away from origo (0,0,0).
Rotation on the other hand... I'll have to investigate further.

Also added BlenderSphere.obj to repository.